+++ src/main/java/com/devwinter/authservice/adapter/input/api/AuthApiController.java
@@ -4,8 +4,8 @@ import com.devwinter.authservice.adapter.input.api.dto.AccessTokenIssuance;
 import com.devwinter.authservice.adapter.input.api.dto.BaseResponse;
 import com.devwinter.authservice.adapter.input.api.dto.MemberLogin;
 import com.devwinter.authservice.adapter.input.api.dto.MemberLogout;
+import com.devwinter.authservice.application.dto.AuthMemberDto;
 import com.devwinter.authservice.application.port.input.AuthMemberUseCase;
-import com.devwinter.authservice.application.port.input.AuthMemberUseCase.AuthMemberDto;
 import com.devwinter.authservice.application.port.input.DeleteRefreshTokenUseCase;
 import com.devwinter.authservice.application.port.input.ExistRefreshTokenValidUseCase;
 import com.devwinter.authservice.application.port.input.GenerateTokenUseCase;
