+++ src/main/java/com/devwinter/authservice/adapter/input/api/dto/MemberValid.java
@@ -1,7 +1,6 @@
 package com.devwinter.authservice.adapter.input.api.dto;
 
-import com.devwinter.authservice.application.port.input.GetMemberValidQuery;
-import com.devwinter.authservice.application.port.input.GetMemberValidQuery.ValidMemberDto;
+import com.devwinter.authservice.application.dto.AuthMemberDto;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -13,9 +12,9 @@ public class MemberValid {
         private Long memberId;
         private String email;
 
-        public static BaseResponse<Response> success(ValidMemberDto memberDto) {
+        public static BaseResponse<Response> success(AuthMemberDto memberDto) {
             return BaseResponse.success("회원 인증에 성공하였습니다.",
-                    new MemberValid.Response(memberDto.memberId(), memberDto.email())
+                    new MemberValid.Response(memberDto.id(), memberDto.email())
             );
         }
     }
