+++ src/main/java/com/devwinter/authservice/application/dto/AuthMemberDto.java
@@ -1,0 +1,9 @@
+package com.devwinter.authservice.application.dto;
+
+import com.devwinter.authservice.domain.Member;
+
+public record AuthMemberDto(Long id, String email) {
+    public static AuthMemberDto of(Member member) {
+        return new AuthMemberDto(member.getId(), member.getEmail());
+    }
+}
