+++ src/main/java/com/devwinter/authservice/application/port/input/GetMemberValidQuery.java
@@ -1,15 +1,9 @@
 package com.devwinter.authservice.application.port.input;
 
-import com.devwinter.authservice.domain.Member;
+import com.devwinter.authservice.application.dto.AuthMemberDto;
 
 public interface GetMemberValidQuery {
 
-    ValidMemberDto valid(Long memberId);
+    AuthMemberDto valid(Long memberId);
 
-    record ValidMemberDto(Long memberId, String email) {
-
-        public static ValidMemberDto of(Member member) {
-            return new ValidMemberDto(member.getId(), member.getEmail());
-        }
-    }
 }
