+++ src/main/java/com/devwinter/authservice/utils/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package com.devwinter.authservice.utils;
 
-import com.devwinter.authservice.application.port.input.AuthMemberUseCase.AuthMemberDto;
+import com.devwinter.authservice.application.dto.AuthMemberDto;
 import com.devwinter.authservice.application.port.input.GenerateTokenUseCase.TokenDto;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
